fix: Error on env tag on yaml.Unmarshaler
Return an error for env tag on a field of a primitive named type
that implements the `yaml.Unmarshaler` interface because for a type
like `type MyString string` that implements `yaml.Unmarshaler`
env unmarshaling will just copy the env value to the underlying
string potentially bypassing any validation logic that might
have been defined in `UnmarshalYAML`. Therefore it's safer to use
`encoding.TextUnmarshaler` instead since it's supported by both
the env and the yaml parsers.