fix: No ErrYAMLMissingConfig behind pointer
Always return ErrYAMLMissingConfig when a non-null YAML value is
assigned to a field of type pointer to struct and a field is missing.
For example, if field type was t *T `yaml:"t"` and T is defined as
type T struct { A int `yaml:"a"`; B int `yaml:"b"` }
then YAML "t:\n  b: 42" must return ErrYAMLMissingConfig because
field "a" wasn't explicitly provided. This must not be the case
though for YAML inputs "t:" or "t: null".